+++ exchangelib/__init__.py
@@ -3,6 +3,7 @@ from .autodiscover import discover
 from .configuration import Configuration
 from .credentials import DELEGATE, IMPERSONATION, Credentials
 from .ewsdatetime import EWSDateTime, EWSTimeZone
+from .restriction import Restriction, Q
 from .services import IdOnly, AllProperties, SHALLOW, DEEP
 from .transport import NTLM, DIGEST, BASIC